+++ src/django_spicy_id/__init__.py
@@ -1,6 +1,7 @@
 from .contrib import monkey_patch_drf
 from .errors import MalformedSpicyIdError, SpicyIdError
 from .fields import (
+    ENCODING_BASE_32,
     ENCODING_BASE_58,
     ENCODING_BASE_62,
     ENCODING_HEX,
@@ -8,6 +9,8 @@ from .fields import (
     SpicyBigAutoField,
     SpicySmallAutoField,
     SpicyUUIDField,
+    TypeIDField,
+    uuid7,
 )
 from .utils import get_url_converter
 
@@ -16,6 +19,8 @@ __all__ = [
     "SpicyAutoField",
     "SpicyBigAutoField",
     "SpicyUUIDField",
+    "TypeIDField",
+    "ENCODING_BASE_32",
     "ENCODING_BASE_58",
     "ENCODING_HEX",
     "ENCODING_BASE_62",
@@ -23,4 +28,5 @@ __all__ = [
     "MalformedSpicyIdError",
     "get_url_converter",
     "monkey_patch_drf",
+    "uuid7",
 ]